# backend/src/main/java/com/tms/backend/tms_backend/dto/PageResponse.java
package com.tms.backend.tms_backend.dto;

import java.util.List;

public record PageResponse<T>(List<T> content, PageMetadata page) {

    public record PageMetadata(int number, int size, long totalElements, int totalPages) {
    }
}