# specs/001-ticket-management-backend/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Ticket Management API
  version: 1.0.0
  description: API-first contract for ticket lifecycle and independent comment management.
servers:
  - url: http://localhost:8080
paths:
  /api/v1/tickets:
    post:
      operationId: createTicket
      summary: Create a ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreateRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
    get:
      operationId: listTickets
      summary: List, filter, and search tickets
      parameters:
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Keyword'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Paginated tickets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketPage'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/v1/tickets/{ticketId}:
    parameters:
      - $ref: '#/components/parameters/TicketId'
    get:
      operationId: getTicket
      summary: Get ticket attributes
      responses:
        '200':
          description: Ticket found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      operationId: updateTicket
      summary: Replace editable ticket fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/tickets/{ticketId}/status:
    parameters:
      - $ref: '#/components/parameters/TicketId'
    patch:
      operationId: updateTicketStatus
      summary: Apply a lifecycle transition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketStatusUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/TransitionError'
  /api/v1/tickets/{ticketId}/comments:
    parameters:
      - $ref: '#/components/parameters/TicketId'
    get:
      operationId: listComments
      summary: List comments for a ticket
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Paginated comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentPage'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      operationId: createComment
      summary: Add a comment to a ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreateRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
components:
  parameters:
    TicketId:
      name: ticketId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    Status:
      name: status
      in: query
      schema:
        $ref: '#/components/schemas/TicketStatus'
    Keyword:
      name: keyword
      in: query
      schema:
        type: string
        maxLength: 150
    Page:
      name: page
      in: query
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 10
    Sort:
      name: sort
      in: query
      schema:
        type: string
        default: createdAt,desc
        example: createdAt,desc
  schemas:
    TicketStatus:
      type: string
      enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED, CANCELLED]
    TicketCreateRequest:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 150
        description:
          type: string
          minLength: 1
        assigneeId:
          type: string
          nullable: true
    TicketUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/TicketCreateRequest'
    TicketStatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/TicketStatus'
    CommentCreateRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
        author:
          type: string
          nullable: true
    TicketResponse:
      type: object
      required: [id, title, description, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/TicketStatus'
        assigneeId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CommentResponse:
      type: object
      required: [id, ticketId, content, author, createdAt]
      properties:
        id:
          type: integer
          format: int64
        ticketId:
          type: integer
          format: int64
        content:
          type: string
        author:
          type: string
        createdAt:
          type: string
          format: date-time
    PageMetadata:
      type: object
      required: [number, size, totalElements, totalPages]
      properties:
        number:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    TicketPage:
      type: object
      required: [content, page]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TicketResponse'
        page:
          $ref: '#/components/schemas/PageMetadata'
    CommentPage:
      type: object
      required: [content, page]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CommentResponse'
        page:
          $ref: '#/components/schemas/PageMetadata'
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    ProblemDetail:
      type: object
      required: [type, title, status, detail, instance, timestamp]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        timestamp:
          type: string
          format: date-time
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
  responses:
    ValidationError:
      description: Invalid request, enum, or pagination input
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFoundError:
      description: Ticket not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    TransitionError:
      description: Invalid or stale lifecycle transition
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'